aggregation to module
